add alot

+++ Assets/EasyMobile/Editor/Common/Utils/FileIO.cs
@@ -56,10 +56,7 @@ namespace EasyMobile.Editor
         {
             path = SlashesToPlatformSeparator(path);
 
-            if (!Directory.Exists(path))
-            {
-                Directory.CreateDirectory(path);
-            }
+            if (!Directory.Exists(path)) Directory.CreateDirectory(path);
         }
 
         /// <summary>
@@ -108,13 +105,10 @@ namespace EasyMobile.Editor
         {
             path = SlashesToPlatformSeparator(path);
 
-            if (!File.Exists(path))
-            {
-                return null;
-            }
+            if (!File.Exists(path)) return null;
 
-            StreamReader sr = new StreamReader(path);
-            string body = sr.ReadToEnd();
+            var sr = new StreamReader(path);
+            var body = sr.ReadToEnd();
             sr.Close();
             return body;
         }
@@ -128,10 +122,7 @@ namespace EasyMobile.Editor
         {
             path = SlashesToPlatformSeparator(path);
 
-            if (!File.Exists(path))
-            {
-                return new string[0];
-            }
+            if (!File.Exists(path)) return new string[0];
 
             return File.ReadAllLines(path);
         }
@@ -144,10 +135,7 @@ namespace EasyMobile.Editor
         {
             path = SlashesToPlatformSeparator(path);
 
-            if (File.Exists(path))
-            {
-                File.Delete(path);
-            }
+            if (File.Exists(path)) File.Delete(path);
         }
 
         /// <summary>
@@ -159,10 +147,7 @@ namespace EasyMobile.Editor
         {
             path = SlashesToPlatformSeparator(path);
 
-            if (Directory.Exists(path))
-            {
-                Directory.Delete(path, recursive);
-            }
+            if (Directory.Exists(path)) Directory.Delete(path, recursive);
         }
 
         /// <summary>
@@ -173,7 +158,7 @@ namespace EasyMobile.Editor
         /// <param name="path">Path.</param>
         public static string ToAbsolutePath(string path)
         {
-            return FileIO.SlashesToPlatformSeparator(Path.Combine(UnityEngine.Application.dataPath, path));
+            return SlashesToPlatformSeparator(Path.Combine(Application.dataPath, path));
         }
 
         /// <summary>
@@ -189,4 +174,3 @@ namespace EasyMobile.Editor
         }
     }
 }
-
